requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.50.0
greenlet==3.0.3
html5lib==1.1
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
passlib==1.7.4
pillow==10.3.0
prompt-toolkit==3.0.43
pycparser==2.22
pydyf==0.9.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
six==1.16.0
SQLAlchemy==2.0.29
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.2
zopfli==0.2.3
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.50.0
greenlet==3.0.3
html5lib==1.1
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
passlib==1.7.4
pillow==10.3.0
prompt-toolkit==3.0.43
pycparser==2.22
pydyf==0.9.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
six==1.16.0
SQLAlchemy==2.0.29
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.2
zopfli==0.2.3
